Split backends logic into classes and subclasses to ease add more backends.
