+++ app/src/backends/Base/index.js
@@ -1,0 +1,6 @@
+import 'abortcontroller-polyfill'; // For GoogleBot
+
+import ByDate from './ByDate';
+import Live from './Live';
+
+export { ByDate, Live };
