--- app/src/rac1/index.js
@@ -1,6 +1,0 @@
-import 'abortcontroller-polyfill'; // For GoogleBot
-
-import Rac1 from './rac1-by-date';
-import Rac1Live from './rac1-directe';
-
-export { Rac1, Rac1Live };
